# web/openapi.yaml
openapi: 3.0.3
info:
  title: ICD Converter API
  version: "1.4.0"
  description: >
    API per la conversione e ricerca di codici ICD-9-CM ↔ ICD-10-IM e codici CIPI (nomenclatura italiana). Supporta lookup puntuale, conversione bidirezionale, espansione gerarchica, ricerca testuale full-text e ricerca per similarità semantica con pipeline avanzata: text enrichment dell'indice, query expansion LLM, multi-query splitting e fusione ibrida RRF (embedding + BM25). Include il grouper MS-DRG CMS FY2026 v43.0 con descrizioni in italiano.
servers:
- url: /api/v1
  description: ICD Converter
...
paths:
...
  /drg/group:
    post:
      tags: [ DRG ]
      summary: Raggruppa un episodio di ricovero in un MS-DRG
      description: |
        Calcola il codice MS-DRG assegnato a un episodio di ricovero ospedaliero a partire da:
        diagnosi principale, diagnosi secondarie (opzionali), procedure ICD-10-PCS (opzionali),
        età, sesso e modalità di dimissione.

        L'algoritmo segue le specifiche ufficiali CMS MS-DRG v43.0 FY2026:
        1. Verifica Pre-MDC (trapianti, ECMO, ustioni estese, ventilazione meccanica prolungata)
        2. Assegnazione MDC dalla diagnosi principale
        3. Partizione Chirurgica / Medica (presenza di procedure O.R.)
        4. Valutazione CC/MCC dalle diagnosi secondarie
        5. Selezione del DRG finale
      operationId: groupDRG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupRequest"
            example:
              principal_dx: "I2109"
              secondary_dxs: [ "J9601", "E1165" ]
              procedures: [ "02703DZ" ]
              age: 67
              sex: "M"
              discharge_status: "01"
      responses:
        "200":
          description: Risultato del raggruppamento DRG
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupResult"
              example:
                drg_code: "246"
                description: "ANGIOPLASTICA CORONARICA PERCUTANEA CON AMI, STENT E CMM"
                mdc: "05"
                mdc_description: "Malattie e Disturbi del Sistema Circolatorio"
                partition: "SURG"
                complication_level: "MCC"
                is_pre_mdc: false
                has_or_procedure: true
                weight: 3.8765
                geometric_los: 4.5
                arithmetic_los: 5.8
                mcc_codes_applied: [ "J9601" ]
                cc_codes_applied: []
                principal_diagnosis:
                  code: "I2109"
                  description: "Infarto miocardico acuto della parete anteriore"
                secondary_diagnoses:
                - code: "J9601"
                  description: "Insufficienza respiratoria acuta con ipossia"
                - code: "E1165"
                  description: "Diabete mellito di tipo 2 con complicazioni circolatorie"
                procedures:
                - code: "02703DZ"
                  description: "Dilatazione dell'arteria coronaria, arteria sinistra anteriore discendente, approccio percutaneo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          description: Codice ICD-10 non trovato o non raggruppabile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "principal diagnosis code not found: XYZ99"

# ── Components ────────────────────────────────────────────────────────────────
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Messaggio di errore
          example: "ICD-9 code not found: 999.9"
...
    DiagnosisDetail:
      type: object
      description: Diagnosi con codice e descrizione risolti.
      properties:
        code:
          type: string
          example: "I2109"
        description:
          type: string
          example: "Infarto miocardico acuto della parete anteriore"
    ProcedureDetail:
      type: object
      description: Procedura con codice e descrizione risolti.
      properties:
        code:
          type: string
          example: "02703DZ"
        description:
          type: string
          example: "Dilatazione dell'arteria coronaria sinistra anteriore discendente, approccio percutaneo"
    GroupRequest:
      type: object
      required: [ principal_dx ]
      description: Input per il grouper MS-DRG.
      properties:
        principal_dx:
          type: string
          description: Diagnosi principale ICD-10-CM (obbligatoria)
          example: "I2109"
        secondary_dxs:
          type: array
          items:
            type: string
          description: Diagnosi secondarie ICD-10-CM (opzionali)
          example: [ "J9601", "E1165" ]
        procedures:
          type: array
          items:
            type: string
          description: Procedure ICD-10-PCS (opzionali)
          example: [ "02703DZ" ]
        age:
          type: integer
          minimum: 0
          maximum: 150
          description: Età del paziente in anni
          example: 67
        sex:
          type: string
          enum: [ M, F, U ]
          description: Sesso del paziente — `M` (maschio), `F` (femmina), `U` (ignoto)
          example: "M"
        discharge_status:
          type: string
          description: Codice modalità di dimissione (01=ordinaria, 02=decesso, ecc.)
          example: "01"
    GroupResult:
      type: object
      description: Risultato del raggruppamento MS-DRG.
      properties:
        drg_code:
          type: string
          description: Codice DRG assegnato
          example: "246"
        description:
          type: string
          description: Descrizione DRG in italiano
          example: "ANGIOPLASTICA CORONARICA PERCUTANEA CON AMI, STENT E CMM"
        mdc:
          type: string
          description: Codice MDC
          example: "05"
        mdc_description:
          type: string
          description: Descrizione MDC in italiano
          example: "Malattie e Disturbi del Sistema Circolatorio"
        partition:
          type: string
          enum: [ SURG, MED ]
          description: Partizione del DRG
          example: "SURG"
        complication_level:
          type: string
          enum: [ MCC, CC, NONE ]
          description: Livello di complicazione/comorbidità applicato — `MCC`, `CC` o `NONE`
          example: "MCC"
        is_pre_mdc:
          type: boolean
          description: Indica se il DRG appartiene alla categoria Pre-MDC
          example: false
        has_or_procedure:
          type: boolean
          description: Indica se è presente almeno una procedura di sala operatoria
          example: true
        weight:
          type: number
          format: float
          description: Peso relativo CMS FY2026
          example: 3.8765
        geometric_los:
          type: number
          format: float
          description: Durata di degenza geometrica media (giorni)
          example: 4.5
        arithmetic_los:
          type: number
          format: float
          description: Durata di degenza aritmetica media (giorni)
          example: 5.8
        mcc_codes_applied:
          type: array
          items:
            type: string
          description: Codici ICD-10-CM di diagnosi secondarie che hanno contribuito come MCC
          example: [ "J9601" ]
        cc_codes_applied:
          type: array
          items:
            type: string
          description: Codici ICD-10-CM di diagnosi secondarie che hanno contribuito come CC
          example: []
        principal_diagnosis:
          $ref: "#/components/schemas/DiagnosisDetail"
        secondary_diagnoses:
          type: array
          items:
            $ref: "#/components/schemas/DiagnosisDetail"
        procedures:
          type: array
          items:
            $ref: "#/components/schemas/ProcedureDetail"
  responses:
...
    BadRequest:
      description: Richiesta non valida (parametro mancante o malformato)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "query parameter 'q' is required"
...